+++ packages/expo/src/winter/runtime.ts
@@ -1,0 +1,5 @@
+Object.defineProperty(globalThis, 'ExpoImportMetaRegistry', {
+  value: require('./ImportMetaRegistry').ImportMetaRegistry,
+  enumerable: true,
+  writable: true,
+});